requirements.txt
# Core LLM APIs
anthropic>=0.39.0
google-generativeai>=0.8.0

# Vector Database and Embeddings
chromadb>=0.4.24
sentence-transformers>=2.7.0

# Document Processing
PyPDF2>=3.0.1
pdfplumber>=0.11.0
pypdf>=4.0.0
python-docx>=1.1.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
tenacity>=8.2.3

# Logging and Monitoring
loguru>=0.7.2

# CLI
click>=8.1.7
rich>=13.7.0

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
